requirements.txt
google-api-python-client==1.7.4
google-auth-httplib2==0.0.3
google-auth==1.5.1
google-cloud-pubsub==0.37.2
cryptography==2.3.1
pyjwt==1.6.4
paho-mqtt==1.3.1
RPi.GPIO==0.6.5
RPLCD==1.2.0
firebase-admin
